Added: ADT_ONLYBOM.
git-svn-id: https://svn.code.sf.net/p/akelpad/codesvn/trunk@3440 24cdac72-59cb-4434-bdd7-86cdcaa6fc6e

+++ akelpad_4/AkelFiles/Plugs/Scripts/Source/IDocument.c
@@ -1588,7 +1588,7 @@ HRESULT STDMETHODCALLTYPE Document_MemRead(IDocument *this, INT_PTR nPointer, DW
   if (dwType == DT_ANSI)
   {
     if (nDataLen == -1)
-      nStringLen=xstrlenA((char *)nPointer);
+      nStringLen=(int)xstrlenA((char *)nPointer);
     else
       nStringLen=nDataLen;
 
@@ -1610,7 +1610,7 @@ HRESULT STDMETHODCALLTYPE Document_MemRead(IDocument *this, INT_PTR nPointer, DW
   else if (dwType == DT_UNICODE)
   {
     if (nDataLen == -1)
-      nStringLen=xstrlenW((wchar_t *)nPointer);
+      nStringLen=(int)xstrlenW((wchar_t *)nPointer);
     else
       nStringLen=nDataLen;
